requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery==2.0.0
google-cloud-bigquery-storage==2.0.0
pandas==1.1.3
pandas-gbq==0.14.0
pyarrow==1.0.1
grpcio==1.32.0